// modules/cip/types/cip.ts
export type CipAutoCompleteType = {
  url?: string;
  name?: string;
  id?: number;
}

export type CipRecentSearchItem = {
  url: string;
  airportName:string;
  flightDate:string;
}